--- TensorLib/tensor_core
@@ -1,5 +1,0 @@
-#include "tensor.cpp"
-#include "tensor_descriptor.cpp"
-#include "tensor_iterator.cpp"
-#include "utils.cpp"
-#include "tensor_formatter.cpp"
