+++ TensorLib/tensor_core.hpp
@@ -1,0 +1,13 @@
+#ifndef TENSORLIB_TENSOR_CORE_H_
+#define TENSORLIB_TENSOR_CORE_H_
+
+#include "tensor_core/tensor.hpp"
+#include "tensor_core/tensor_descriptor.hpp"
+#include "tensor_core/tensor_iterator.hpp"
+#include "tensor_core/tensor_formatter.hpp"
+#include "tensor_core/tensor_print.hpp"
+#include "tensor_core/slice.hpp"
+#include "tensor_core/range.hpp"
+#include "tensor_core/utils.hpp"
+
+#endif // TENSORLIB_TENSOR_CORE_H_
